+++ esp32c3-embassy/src/sensor.rs
@@ -18,7 +18,7 @@ use embassy_sync::blocking_mutex::raw::NoopRawMutex;
 use embassy_sync::channel::Sender;
 
 use esp_hal::i2c::Error as I2cError;
-use esp_hal::i2c::I2C;
+use esp_hal::i2c::I2c;
 use esp_hal::peripherals::I2C0;
 use esp_hal::rng::Rng;
 use esp_hal::Async;
@@ -41,7 +41,7 @@ const WARMUP_INTERVAL: Duration = Duration::from_millis(10);
 /// Task for sampling sensor
 #[embassy_executor::task]
 pub async fn sample_task(
-    i2c: I2C<'static, I2C0, Async>,
+    i2c: I2c<'static, I2C0, Async>,
     mut rng: Rng,
     sender: Sender<'static, NoopRawMutex, Reading, 3>,
     clock: Clock,
@@ -73,7 +73,7 @@ pub async fn sample_task(
 
 /// Sample sensor and send reading to receiver
 async fn sample_and_send(
-    sensor: &mut AsyncBme280<I2C<'static, I2C0, Async>, Delay>,
+    sensor: &mut AsyncBme280<I2c<'static, I2C0, Async>, Delay>,
     rng: &mut Rng,
     sender: &Sender<'static, NoopRawMutex, Reading, 3>,
     clock: &Clock,
@@ -102,7 +102,7 @@ async fn sample_and_send(
 
 /// Initialize sensor
 async fn initialize(
-    bme280: &mut AsyncBme280<I2C<'static, I2C0, Async>, Delay>,
+    bme280: &mut AsyncBme280<I2c<'static, I2C0, Async>, Delay>,
 ) -> Result<(), I2cError> {
     info!("Initialize");
     bme280.init().await?;
@@ -124,13 +124,13 @@ async fn initialize(
 #[derive(Debug)]
 enum SensorError {
     /// Error from clock
-    Clock(#[allow(unused)] ClockError),
+    Clock(#[expect(unused, reason = "Never read directly")] ClockError),
 
     /// Error from domain
-    Domain(#[allow(unused)] DomainError),
+    Domain(DomainError),
 
     /// Error from I²C bus
-    I2c(#[allow(unused)] I2cError),
+    I2c(#[expect(unused, reason = "Never read directly")] I2cError),
 }
 
 impl From<ClockError> for SensorError {