+++ flake.nix
@@ -1,0 +1,50 @@
+{
+ inputs = {
+    nixpkgs.url = "nixpkgs/nixos-unstable";
+    flake-utils.url = "github:numtide/flake-utils";
+    naersk = {
+      url = "github:nix-community/naersk";
+      inputs.nixpkgs.follows = "nixpkgs";
+    };
+    fenix = {
+      url = "github:nix-community/fenix";
+      inputs.nixpkgs.follows = "nixpkgs";
+    };
+  };
+
+  outputs = { self, nixpkgs, flake-utils, naersk, fenix }:
+    flake-utils.lib.eachDefaultSystem (system:
+      let
+        pkgs = (import nixpkgs) {
+          inherit system;
+        };
+
+        toolchain = with fenix.packages.${system}; fromToolchainFile {
+          file = ./rust-toolchain.toml; # alternatively, dir = ./.;
+          sha256 = "sha256-yMuSb5eQPO/bHv+Bcf/US8LVMbf/G/0MSfiPwBhiPpk=";
+        };
+
+        target = "riscv32imc-unknown-none-elf";
+
+      in {
+        packages.default =
+          (naersk.lib.${system}.override {
+            cargo = toolchain;
+            rustc = toolchain;
+          }).buildPackage {
+            src = ./.;
+            CARGO_BUILD_TARGET = target;
+            WIFI_SSID = "xxx";
+            WIFI_PASSWORD = "xxx";
+          };
+
+        # For `nix develop` or `direnv allow`:
+        devShell = pkgs.mkShell {
+          buildInputs = with pkgs; [
+            toolchain
+            rust-analyzer
+          ];
+        };
+      }
+    );
+}