+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2024-11-05
+
+### Added
+
+- Nix flake (contributed by [MaxKiv](https://github.com/MaxKiv/))
+
+### Changed
+
+- Update `embassy-executor` to version 0.6
+- Update `esp-hal` to version 0.21
+- Update `esp-hal-embassy` to version 0.4
+- Update `esp-backtrace` to version 0.14
+- Update `esp-println` to version 0.12
+- Update `esp-wifi` to version 0.10
+- Add feature `esp-alloc` to `esp-wifi`
+- Add `esp-alloc` to version 0.5
+- Update `bme208-rs` to version 0.3
+- Update `uom` to version 0.36
+
+
 ## [0.3.0] - 2024-09-09
 
 ### Changed
@@ -41,3 +61,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.1.0]: https://gitlab.com/claudiomattera/esp32c3-embassy/-/tags/0.1.0
 [0.2.0]: https://gitlab.com/claudiomattera/esp32c3-embassy/-/tags/0.2.0
 [0.3.0]: https://gitlab.com/claudiomattera/esp32c3-embassy/-/tags/0.3.0
+[0.4.0]: https://gitlab.com/claudiomattera/esp32c3-embassy/-/tags/0.4.0